+++ app/src/main/java/com/sebastianvm/musicplayer/repository/playlist/PlaylistRepository.kt
@@ -4,7 +4,7 @@ import com.sebastianvm.musicplayer.database.entities.Playlist
 import kotlinx.coroutines.flow.Flow
 
 interface PlaylistRepository {
-    fun getPlaylistsCount(): Flow<Long>
+    fun getPlaylistsCount(): Flow<Int>
     fun getPlaylists(): Flow<List<Playlist>>
     suspend fun createPlaylist(playlistName: String)
     suspend fun deletePlaylist(playlistName: String)
